+++ notification/src/main/java/com/microservice/erp/domain/dto/NoticeDto.java
@@ -3,11 +3,13 @@ package com.microservice.erp.domain.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Setter
 @Getter
 public class NoticeDto {
     //region private variables
-    private Long noticeConfigurationId;
+    private BigInteger noticeConfigurationId;
     private String year;
     private String noticeName;
     //endregion
