+++ notification/src/main/java/com/microservice/erp/services/iServices/IReadNoticeConfigurationService.java
@@ -2,11 +2,12 @@ package com.microservice.erp.services.iServices;
 
 import com.microservice.erp.domain.entities.NoticeConfiguration;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IReadNoticeConfigurationService {
 
-    NoticeConfiguration getAllNoticeConfigurationById(Long id);
+    NoticeConfiguration getAllNoticeConfigurationById(BigInteger id);
 
     List<NoticeConfiguration> getAllNoticeConfigurationList();
 }
