+++ training-management/src/main/java/com/microservice/erp/controllers/rest/ReasonController.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -24,7 +25,7 @@ public class ReasonController {
 
 
     @GetMapping("/getAllReasonById")
-    public Reason getAllReasonById(@RequestParam("id")Long id) {
+    public Reason getAllReasonById(@RequestParam("id") BigInteger id) {
         return readService.getAllReasonById(id);
     }
