+++ training-management/src/main/java/com/microservice/erp/domain/entities/MedicalQuestionCategory.java
@@ -1,13 +1,14 @@
 package com.microservice.erp.domain.entities;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.Basic;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.validation.constraints.NotNull;
-import java.util.Objects;
+import java.math.BigInteger;
 
 /**
  * @author Rajib Kumer Ghosh
@@ -16,16 +17,16 @@ import java.util.Objects;
 @AllArgsConstructor
 @Entity(name = "tms_medical_question_category")
 @AttributeOverride(name = "id", column = @Column(name = "medical_question_category_id"))
-public class MedicalQuestionCategory extends Auditable<Long, Long> {
+public class MedicalQuestionCategory extends Auditable<BigInteger, Long> {
 
     @Basic(optional = false)
     @NotNull(message = "Category name cannot be null.")
-    @Column(name = "category_name",columnDefinition = "varchar(255)")
+    @Column(name = "category_name", columnDefinition = "varchar(255)")
     private String categoryName;
 
     @NotNull
     @Basic(optional = false)
-    @Column(name = "status",columnDefinition = "char(1)")
+    @Column(name = "status", columnDefinition = "char(1)")
     private String status;
 
     public String getCategoryName() {
