+++ training-management/src/main/java/com/microservice/erp/domain/repository/IDzongkhagHospitalMappingRepository.java
@@ -4,6 +4,7 @@ import com.microservice.erp.domain.entities.DzongkhagHospitalMapping;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -11,7 +12,7 @@ import java.util.List;
  */
 
 @Repository
-public interface IDzongkhagHospitalMappingRepository extends JpaRepository<DzongkhagHospitalMapping, Long> {
+public interface IDzongkhagHospitalMappingRepository extends JpaRepository<DzongkhagHospitalMapping, BigInteger> {
     DzongkhagHospitalMapping findByDzongkhagId(Integer id);
 
     List<DzongkhagHospitalMapping> findAllByStatus(String status);
