+++ training-management/src/main/java/com/microservice/erp/domain/repository/IReasonRepository.java
@@ -5,10 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface IReasonRepository extends JpaRepository<Reason, Long> {
+public interface IReasonRepository extends JpaRepository<Reason, BigInteger> {
 
     List<Reason> findAllByStatus(String status);
 }
