+++ training-management/src/main/java/com/microservice/erp/services/iServices/IReadDzongkhagHospitalMappingService.java
@@ -3,13 +3,14 @@ package com.microservice.erp.services.iServices;
 import com.microservice.erp.domain.entities.DzongkhagHospitalMapping;
 import org.springframework.http.ResponseEntity;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IReadDzongkhagHospitalMappingService {
 
     List<DzongkhagHospitalMapping> getAllDzongkhagHospitalList();
 
-    DzongkhagHospitalMapping getAllDzongkhagHospitalById(Long id);
+    DzongkhagHospitalMapping getAllDzongkhagHospitalById(BigInteger id);
 
     List<DzongkhagHospitalMapping> getAllDzongkhagHosByStatus(String status);
