+++ training-management/src/main/java/com/microservice/erp/services/iServices/IReadTrainingAcademyCapacityService.java
@@ -2,10 +2,11 @@ package com.microservice.erp.services.iServices;
 
 import com.microservice.erp.domain.entities.TrainingAcademyCapacity;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IReadTrainingAcademyCapacityService {
     List<TrainingAcademyCapacity> getAllTrainingAcaCapList();
 
-    TrainingAcademyCapacity getAllTrainingAcaCapById(Long id);
+    TrainingAcademyCapacity getAllTrainingAcaCapById(BigInteger id);
 }
